Move CanonicalDecomposition, CanonicalComposition, and CanonicalCombiningClassMap to a module
